Update ItemController to have authorization

+++ WishList/Controllers/ItemController.cs
@@ -1,21 +1,29 @@
 ﻿using System.Linq;
+using Microsoft.AspNetCore.Authorization;
+using Microsoft.AspNetCore.Identity;
 using Microsoft.AspNetCore.Mvc;
 using WishList.Data;
+using WishList.Models;
 
 namespace WishList.Controllers
 {
+    [Authorize]
     public class ItemController : Controller
     {
+        private readonly UserManager<ApplicationUser> _userManager;
         private readonly ApplicationDbContext _context;
 
-        public ItemController(ApplicationDbContext context)
+        public ItemController(ApplicationDbContext context, UserManager<ApplicationUser> userManager)
         {
+            _userManager=userManager;
             _context = context;
         }
 
         public IActionResult Index()
         {
-            var model = _context.Items.ToList();
+            var currentUser=_userManager.GetUserAsync(HttpContext.User).Result;
+            
+            var model = _context.Items.ToList().Where(i=>i.User.Id == currentUser.Id);
 
             return View("Index", model);
         }
@@ -36,7 +44,10 @@ namespace WishList.Controllers
 
         public IActionResult Delete(int id)
         {
+        
             var item = _context.Items.FirstOrDefault(e => e.Id == id);
+            if(item.User.Id!=_userManager.GetUserAsync(HttpContext.User).Result.Id)
+                return Unauthorized();
             _context.Items.Remove(item);
             _context.SaveChanges();
             return RedirectToAction("Index");
